sql-agent-system/requirements.txt
# Core LLM & Orchestration
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-google-genai>=0.0.1
langchain_core>=0.1.0
langchain_community>=0.0.1
langgraph>=0.0.1
langchain-huggingface>=0.0.1
# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Schema & Validation
pydantic>=2.0.0

# Monitoring & Observability
langfuse>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# MCP Server (Model Context Protocol)
mcp>=0.9.0

# UI Framework
gradio>=4.0.0

# REST API
fastapi>=0.104.0
uvicorn>=0.24.0

# Optional: Advanced Features (uncomment as needed)
# redis>=5.0.0  # For caching
# prometheus-client>=0.19.0  # For metrics export
# sentry-sdk>=1.38.0  # For error tracking
